cmd/server: better ui base path defaults, move to server.go from mockui.go to make it clear it affects non-mock code too
