+++ pkg/config/config.go
@@ -17,8 +17,6 @@ type ServerConfig struct {
 	DisableTLS                 bool
 	TLSCert                    tls.Certificate
 	WordnetPath                string
-	FuranAddrs                 []string
-	EnableFuran2               bool
 	Furan2Addr                 string
 	Furan2APIKey               string
 	Furan2SkipVerifyTLS        bool
